Refactor routing: implement react-router for navigation and restructure app layout; remove old Home component
