+++ src/App.jsx
@@ -1,12 +1,21 @@
+import {BrowserRouter, Route, Routes} from 'react-router';
 import './App.css';
-import Home from './components/Home';
+import Home from './views/Home';
+import Layout from './components/Layout';
+import Profile from './views/Profile';
+import Upload from './views/Upload';
 
 function App() {
   return (
-    <>
-      <h1>My App</h1>
-      <Home />
-    </>
+    <BrowserRouter>
+      <Routes>
+        <Route element={<Layout />}>
+          <Route path="/" element={<Home />} />
+          <Route path="/profile" element={<Profile />} />
+          <Route path="/upload" element={<Upload />} />
+        </Route>
+      </Routes>
+    </BrowserRouter>
   );
 }
