+++ src/views/Home.jsx
@@ -1,6 +1,6 @@
 import {useState} from 'react';
-import MediaRow from './MediaRow';
-import SingleView from './SingleView';
+import MediaRow from '../components/MediaRow';
+import SingleView from '../components/SingleView';
 
 const mediaArray = [
   {
