+++ src/views/Profile.jsx
@@ -1,0 +1,5 @@
+const Profile = () => {
+  return <div>Profile</div>;
+};
+
+export default Profile;
