+++ src/views/Single.jsx
@@ -1,0 +1,5 @@
+const Single = () => {
+  return <div>Single</div>;
+};
+
+export default Single;
