+++ src/views/Upload.jsx
@@ -1,0 +1,5 @@
+const Upload = () => {
+  return <div>Upload</div>;
+};
+
+export default Upload;
